# web/src/lib/schemas/openapi-v3.yml
openapi: 3.1.0
info:
  title: Talaria V3 entity projections
  version: 1.0.0
paths:
...
  /api/v1/claims/{claim_id}/comments:
    get:
      summary: List Agora comments for a claim
      description: |
        Public. Roots newest first. Replies oldest first, max 20 per root.
        Deleted comments keep their place; body is omitted. Off-chain only.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: cursor
          in: query
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 50}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Comment'}
                  next_cursor: {type: [string, 'null']}
        '400': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
    post:
      summary: Create a root comment or a single-level reply
      description: Session required. parent_comment_id must be a root of this claim.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: {type: string, minLength: 1, maxLength: 4000}
                parent_comment_id: {type: [string, 'null'], format: uuid}
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                required: [comment]
                properties:
                  comment: {$ref: '#/components/schemas/Comment'}
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
  /api/v1/comments/{comment_id}:
    patch:
      summary: Edit own active comment
      description: Foreign ownership returns comment_not_found.
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: {type: string, minLength: 1, maxLength: 4000}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment: {$ref: '#/components/schemas/Comment'}
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
    delete:
      summary: Soft-delete own comment
      description: status=deleted; replies remain; body omitted from public DTO.
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment: {$ref: '#/components/schemas/Comment'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
  /api/v1/comments/{comment_id}/reactions:
    post:
      summary: Add a semantic reaction (Talaria-only, not an Intuition signal)
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reaction_type]
              properties:
                reaction_type: {$ref: '#/components/schemas/CommentReactionType'}
      responses:
        '200':
          description: Already present (idempotent)
        '201':
          description: Created
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
  /api/v1/comments/{comment_id}/reactions/{reaction_type}:
    delete:
      summary: Remove own semantic reaction
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: reaction_type
          in: path
          required: true
          schema: {$ref: '#/components/schemas/CommentReactionType'}
      responses:
        '200':
          description: Comment with updated counts
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
components:
...
  responses:
...
    StructuredError:
      description: Code + message, never raw SQL
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code:
                    type: string
                    enum:
                      - unauthenticated
                      - invalid_action
                      - invalid_target_type
                      - invalid_action_for_target
                      - target_not_found
                      - interaction_not_found
                      - invalid_visibility
                      - batch_too_large
                      - database_error
                      - invalid_address
                      - conflict
                      - expired_challenge
                      - invalid_signature
                  message: {type: string}
  schemas:
...
    CommentReactionType:
      type: string
      enum: [relevant, well_sourced, interesting, needs_nuance, disagree]
    Comment:
      type: object
      required: [id, claim_id, author, status, created_at, reactions, my_reactions]
      properties:
        id: {type: string, format: uuid}
        claim_id: {type: string, format: uuid}
        author:
          type: object
          required: [id, display_name]
          properties:
            id: {type: string, format: uuid}
            display_name: {type: [string, 'null']}
        body: {type: [string, 'null']}
        status: {type: string, enum: [active, deleted]}
        edited_at: {type: [string, 'null'], format: date-time}
        created_at: {type: string, format: date-time}
        reactions:
          type: object
          additionalProperties: {type: integer}
        my_reactions:
          type: array
          items: {$ref: '#/components/schemas/CommentReactionType'}
        replies:
          type: array
          items: {$ref: '#/components/schemas/Comment'}
        reply_count: {type: integer}
...